Move generic transformation code into a dedicated transformation.py
